Feat : 댓글 커뮤니티 boardId랑 연결하기
